Renamed escrow to payment in file names and paths
