+++ Cargo.toml
@@ -4,7 +4,7 @@ panic = 'unwind'
 [workspace]
 members = [
     'node/parachain',
-    'pallets/escrow',
+    'pallets/payment',
     'primitives',
     'runtime'
 ]
